Delete action improved

+++ ProductsAPI/Service/OrderService.cs
@@ -77,6 +77,12 @@ namespace ProductsApi.Service
         public async Task<bool> DeleteAsync(int Id)
         {
             var accountToDelete = await _orderRepository.GetByIdAsync(Id);
+
+            if (accountToDelete == null)
+            {
+                return false;
+            }
+
             return await _orderRepository.DeleteAsync(accountToDelete);
         }
     }
